If the build is explicitly triggered for a tag, fetch the message from that tag.
Previously, we were always fetching the tag message from the most recent tag
associated with the checked-out commit, even if that tag did not match the
configured refspec.

But now the first tag mentioned in the "Checking out" console output will be the
tag from which we fetch and export the tag messasge.