Modernize plugin
With the upgrade to the latest parent pom, e.g. more meta-info and less
jar files are packaged in the .hpi, more checks are performed in the build.

Because Java 7 is end-of-life since a while, upgrading to Java 8 and
one of the first Jenkins LTS versions that requires it.

Upgrading git dependency to be more up-to-date and that less other
dependency versions have to be declared to comply with the
maven-enforcer plugin.